+++ main.py
@@ -34,19 +34,16 @@ DATASET = 'cifar10'
 # NETD_MNIST = './samples/mnist/netD_epoch_24.pth'
 
 
-
-NUM_EPOCHS = 25 if not isDebug else 10
-D_ITERS = 5 if not isDebug else 0                           # Pre-Train D Number of D iterations per G iteration
-MAX_ITERS_PER_EPOCH = 100 if not isDebug else 20            # Usually it's all images in trainloader approx. 469
-
-BATCH_SIZE = 128 if not isDebug else 64
+NUM_EPOCHS = 25
+BATCH_SIZE = 128
 IMG_SIZE = 64
 IMG_CHANNELS = 3
 NGF = BATCH_SIZE * 2
 NDF = BATCH_SIZE * 2
-
 LR_D = 0.00005
 LR_G = 0.00005
+D_ITERS = 5             # Number of D iterations per G iteration
+
 
 
 def getOptimizers(opt, netG, netD):
@@ -91,8 +88,6 @@ def __getGenerator(opt, ngpu, nz, ngf, ndf, nc, n_extra_layers):
     elif opt.mlp_G:
         if isDebug: print("Using MLP_G for Generator")
         netG = mlp.MLP_G(opt.imageSize, nz, nc, ngf, ngpu)
-    elif opt.caps_D:
-        netG = dcgan.DCGAN_G(32, nz, nc, ngf, ngpu, n_extra_layers, bias=False, opt=opt)
     else:
         if isDebug: print("Using DCGAN_G for Generator")
         netG = dcgan.DCGAN_G(opt.imageSize, nz, nc, ngf, ngpu, n_extra_layers, bias=False)
@@ -110,7 +105,7 @@ def __getDiscriminator(opt, ngpu, nz, ngf, ndf, nc, n_extra_layers):
         netD = mlp.MLP_D(opt.imageSize, nz, nc, ndf, ngpu)
     elif opt.caps_D:
         if isDebug: print("Using CapsNet for Discriminator/Critic")
-        netD =capsgan.CapsNet(opt)
+        netD =capsgan.CapsNet(opt=opt)
     else:
         if isDebug: print("Using DCGAN_D for Discriminator/Critic")
         netD = dcgan.DCGAN_D(opt.imageSize, nz, nc, ndf, ngpu, n_extra_layers, False)
@@ -134,6 +129,7 @@ def weights_init(m):
         m.weight.data.normal_(1.0, 0.02)
         m.bias.data.fill_(0)
 
+
 def main(opt):
     cuda = opt.cuda; visualize = opt.visualize
     print(f"cuda = {cuda}, visualize = {opt.visualize}")
@@ -155,169 +151,18 @@ def main(opt):
     if USE_CUDA and not opt.cuda:
         eprint("WARNING: CUDA device available, please run with CUDA")
 
+    # dataset = __getDataSet(opt)
     dataset = getDataSet(opt)
     assert dataset
-    dataloader = torch.utils.data.DataLoader(dataset,
-                                             batch_size=opt.batchSize,
-                                             shuffle=True,
-                                             num_workers=int(opt.workers))
-
     if opt.dataset == 'mnist':
         train_dataset = dataset.train_dataset
         test_dataset = dataset.test_dataset
         dataset = dataset.full_dataset
-        dataloader = torch.utils.data.DataLoader(train_dataset,
-                                                 batch_size=opt.batchSize,
-                                                 shuffle=True,
-                                                 num_workers=int(opt.workers))
-    print(f"len(dataloader) = {len(dataloader)}")
-    nz = int(opt.nz)
-    nc = int(opt.nc)
 
-    input = torch.FloatTensor(opt.batchSize, nc, opt.imageSize, opt.imageSize)
-    noise = torch.FloatTensor(opt.batchSize, nz, 1, 1)
-    fixed_noise = torch.FloatTensor(opt.batchSize, nz, 1, 1).normal_(0, 1)
-    one = torch.FloatTensor([1])
-    mone = one * -1
-
-    ## Get Networks
-    netG, netD = getNetworks(opt)
-    if opt.cuda:
-        if isDebug: print("Using CUDA")
-        netD.cuda()
-        netG.cuda()
-        input = input.cuda()
-        one, mone = one.cuda(), mone.cuda()
-        noise, fixed_noise = noise.cuda(), fixed_noise.cuda()
-
-    ## Setup Optimizers
-    optimizerD = optim.Adam(netD.parameters(), lr=opt.lrD, betas=(0.5, 0.999))
-    optimizerG = optim.Adam(netG.parameters(), lr=opt.lrG, betas=(0.5, 0.999))
-
-    print(f"len(dataloader) = {len(dataloader)}")
-    for epoch in tqdm(range(opt.niter)):
-        data_iter = iter(dataloader)
-        i = 0
-        while i < len(dataloader) and i < MAX_ITERS_PER_EPOCH:
-            # Train Discriminator
-            netD.zero_grad()
-            ## Get Real and Fake Data for iter ##
-            data = data_iter.next()
-            real_cpu, _ = data
-            batch_size = real_cpu.size(0)
-            y_real = Variable(torch.ones(batch_size))
-            y_fake = Variable(torch.zeros(batch_size))
-
-            # in case our last batch was the tail batch of the dataloader,
-            # make sure we feed a full batch of noise
-            z = torch.randn((opt.batchSize, nz)).view(-1, nz, 1, 1)
-            z = Variable(z.cuda()) if opt.cuda else Variable(z)
-            Gz = Variable(netG(z).data, volatile=True)
-            if opt.cuda:
-                real_cpu = real_cpu.cuda()
-                # real_data = real_data.cuda()
-                y_real = y_real.cuda()
-                y_fake = y_fake.cuda()
-
-            X = Variable(input.resize_as_(real_cpu).copy_(real_cpu))
-            V, reconstructions, masked = netD(X)
-            netD_real_loss = netD.loss(data=X, x=V, target=y_real, reconstructions=reconstructions)
-
-            Vz, reconstructions_z, masked_z = netD(Gz)
-            netD_fake_loss = netD.loss(data=Gz, x=Vz, target=y_fake, reconstructions=reconstructions_z)
-
-            # netD_train_loss = netD_real_loss + netD_fake_loss
-            netD_train_loss = netD_real_loss -  netD_fake_loss
-            # print(f"\n\tepoch_{epoch}_batch_{i}_D_real_loss = {netD_real_loss}")
-            # print(f"\tepoch_{epoch}_batch_{i}_D_fake_loss = {netD_fake_loss}\n")
-            print(f"\n\tepoch_{epoch}_batch_{i}_D_train_loss = {netD_train_loss}\n")
-
-            netD_train_loss.backward()
-            optimizerD.step()
-
-            ## Train Generator ##
-            netG.zero_grad()
-            # optimizerG.zero_grad()
-            # netD.eval()
-            z = torch.randn((opt.batchSize, nz)).view(-1, nz, 1, 1)
-            z = Variable(z.cuda()) if opt.cuda else Variable(z)
-            Gz = Variable(netG(z).data, volatile=True)
-            y_real = Variable(torch.ones(batch_size))
-            if opt.cuda:
-                y_real = y_real.cuda()
-
-            Vz, reconstructions_z, masked_z = netD(Gz)
-            netG_train_loss = netD.loss(data=Gz, x=Vz, target=y_real, reconstructions=reconstructions_z)
-            print(f"\n\tepoch_{epoch}_batch_{i}_G_train_loss = {netG_train_loss}\n")
-            netG_train_loss.backward()
-            optimizerG.step()
-
-
-            if visualize:
-                netD_loss_logger.log(epoch, netD_train_loss)
-                netG_loss_logger.log(epoch, netG_train_loss)
-
-            # if gen_iterations % 10 == 0 or ((gen_iterations % 100 == 0) and (opt.dataset == 'mnist')):
-            if i % 10 == 0 or ((i % 100 == 0) and (opt.dataset == 'mnist')):
-                real_cpu = real_cpu.mul(0.5).add(0.5)       #de-normalizing mnist to get real sample x_real = mu + sigma.z
-                vutils.save_image(real_cpu, '{0}/{1}/real_samples.png'.format(opt.experiment, opt.dataset))
-
-                fake = netG(Variable(fixed_noise, volatile=True))   # Get a sample from random data from the generator
-                fake.data = fake.data.mul(0.5).add(0.5)
-                vutils.save_image(fake.data, '{0}/{1}/fake_samples_epoch_{2}_batch_{3}.png'.format(opt.experiment, opt.dataset, epoch, i))
-
-
-            i += 1
-
-
-        # do checkpointing
-        if opt.niter > 25:
-            if epoch % 10 == 0:
-                torch.save(netG.state_dict(), '{0}/{1}/netG_epoch_{2}.pth'.format(opt.experiment, opt.dataset, epoch))
-                torch.save(netD.state_dict(), '{0}/{1}/netD_epoch_{2}.pth'.format(opt.experiment, opt.dataset, epoch))
-        else:
-            torch.save(netG.state_dict(), '{0}/{1}/netG_epoch_{2}.pth'.format(opt.experiment, opt.dataset, epoch))
-            torch.save(netD.state_dict(), '{0}/{1}/netD_epoch_{2}.pth'.format(opt.experiment, opt.dataset, epoch))
-
-
-
-def main2(opt):
-    cuda = opt.cuda; visualize = opt.visualize
-    print(f"cuda = {cuda}, visualize = {opt.visualize}")
-    if visualize:
-        netD_loss_logger = VisdomPlotLogger('line', opts={'title': 'Discriminator (NetD) Loss'})
-        netG_loss_logger = VisdomPlotLogger('line', opts={'title': 'Generator (NetG) Loss'})
-
-    cudnn.benchmark = True
-    opt.manualSeed = random.randint(1, 10000)  # fix seed
-    print("Random Seed: ", opt.manualSeed)
-    random.seed(opt.manualSeed)
-    torch.manual_seed(opt.manualSeed)
-
-    ## Path to generative samples storage
-    if opt.experiment is None:
-        opt.experiment = 'samples'
-    os.system('mkdir {0}'.format(opt.experiment))
-
-    if USE_CUDA and not opt.cuda:
-        eprint("WARNING: CUDA device available, please run with CUDA")
-
-    dataset = getDataSet(opt)
-    assert dataset
     dataloader = torch.utils.data.DataLoader(dataset,
                                              batch_size=opt.batchSize,
                                              shuffle=True,
                                              num_workers=int(opt.workers))
-
-    if opt.dataset == 'mnist':
-        train_dataset = dataset.train_dataset
-        test_dataset = dataset.test_dataset
-        dataset = dataset.full_dataset
-        dataloader = torch.utils.data.DataLoader(train_dataset,
-                                                 batch_size=opt.batchSize,
-                                                 shuffle=True,
-                                                 num_workers=int(opt.workers))
-    print(f"len(dataloader) = {len(dataloader)}")
     nz = int(opt.nz)
     nc = int(opt.nc)
 
@@ -341,153 +186,85 @@ def main2(opt):
     optimizerG, optimizerD = getOptimizers(opt, netG, netD)
 
     gen_iterations = 0
-    print(f"len(dataloader) = {len(dataloader)}")
     for epoch in tqdm(range(opt.niter)):
         data_iter = iter(dataloader)
         i = 0
-        # while i < len(dataloader):
-
-        while i < MAX_ITERS_PER_EPOCH and i < len(dataloader):
-            i += 1
+        while i < len(dataloader):
             ############################
-            # (1) Train D network (netD)
+            # (1) Update D network
             ###########################
-            # for p in netD.parameters():  # reset requires_grad
-            #     p.requires_grad = True  # they are set to False below in netG update
+            for p in netD.parameters():  # reset requires_grad
+                p.requires_grad = True  # they are set to False below in netG update
 
             # train the discriminator Diters times
-            # if gen_iterations < 25 or gen_iterations % 500 == 0:
-            #     Diters = 100
-            # else:
-            #     Diters = opt.Diters
-            Diters = opt.Diters
-            shouldClampNetworkParams = False
-            print(f"Starting Pre-Training Discriminator of {Diters} iterations ... ")
+            if gen_iterations < 25 or gen_iterations % 500 == 0:
+                Diters = 100
+            else:
+                Diters = opt.Diters
             j = 0
-            # while j < Diters and i < len(dataloader):
-            '''
-            while j < Diters and i < MAX_ITERS_PER_EPOCH:
+            while j < Diters and i < len(dataloader):
                 j += 1
-                if shouldClampNetworkParams:
-                    # clamp parameters to a cube
-                    for p in netD.parameters():
-                        p.data.clamp_(opt.clamp_lower, opt.clamp_upper)
+
+                # clamp parameters to a cube
+                for p in netD.parameters():
+                    p.data.clamp_(opt.clamp_lower, opt.clamp_upper)
 
                 data = data_iter.next()
-                print("-" * 5 + "Train with real " + "-" * 5)
-                netD.zero_grad()
+                i += 1
+
+                # train with real
                 real_cpu, _ = data
+                netD.zero_grad()
                 batch_size = real_cpu.size(0)
-                y_real = Variable(torch.ones(batch_size))
-                y_fake = Variable(torch.zeros(batch_size))
+
                 if opt.cuda:
                     real_cpu = real_cpu.cuda()
-                    # real_data = real_data.cuda()
-                    y_real = Variable(y_real.cuda())
-                    y_fake = Variable(y_fake.cuda())
-
                 input.resize_as_(real_cpu).copy_(real_cpu)
-                real_data = Variable(input)  # x_
+                inputv = Variable(input)
 
-                optimizerD.zero_grad()                          # clear accum. grads from prev batch
-                V, reconstructions, masked = netD(real_data)
-                netD_real_loss = netD.loss(data=real_data, x=V, target=y_real, reconstructions=reconstructions)
-                print(f"\n\tepoch_{epoch}_batch_{j}_D_real_loss = {netD_real_loss}\n")
-                z = torch.randn((batch_size, nz)).view(-1, nz, 1, 1)    # e.g. 128, 100, 1, 1
-                z = Variable(z.cuda()) if opt.cuda else Variable(z)
+                errD_real = netD(inputv)
+                errD_real.backward(one)
 
-                print("-" * 5 + "Train with fake " + "-" * 5)
-                Gz = netG(z)
-                Gz = Variable(Gz.data, volatile=True)
-                #Gz = Variable(netG(z).data)
-                # Fixed: here Gz (output of G is 128, 1, 32, 32) -> need (128, 1, 28, 28) for feeding to caps_D
-
-                Vz, reconstructions_z, masked_z = netD(Gz)
-                netD_fake_loss = netD.loss(data=Gz, x=Vz, target=y_fake, reconstructions=reconstructions_z)
-                print(f"\tepoch_{epoch}_batch_{j}_D_fake_loss = {netD_fake_loss}")
-
-                netD_train_loss = netD_real_loss + netD_fake_loss           ## TRAINING LOSS
-                print(f"\tepoch_{epoch}_batch_{j}_D_train_loss = {netD_train_loss}")
-                netD_train_loss.backward()
+                # train with fake
+                noise.resize_(opt.batchSize, nz, 1, 1).normal_(0, 1)
+                noisev = Variable(noise, volatile=True)  # totally freeze netG
+                fake = Variable(netG(noisev).data)
+                inputv = fake
+                errD_fake = netD(inputv)
+                errD_fake.backward(mone)
+                errD = errD_real - errD_fake
                 optimizerD.step()
-            '''
+
             ############################
-            # (2) Update Adversarial D, G network
+            # (2) Update G network
             ###########################
-            print(f"Starting Training Discriminator, Generator for {MAX_ITERS_PER_EPOCH-j} iterations ... ")
-
-            ## Reset Grad ##
-            netD.zero_grad()
+            for p in netD.parameters():
+                p.requires_grad = False  # to avoid computation
             netG.zero_grad()
-            optimizerD.zero_grad()
-            optimizerG.zero_grad()
-            # netD.eval()
-
-            ## Get Real and Fake Data for iter ##
-            data = data_iter.next()
-            real_cpu, _ = data
-            batch_size = real_cpu.size(0)
-            y_real = Variable(torch.ones(batch_size))
-
             # in case our last batch was the tail batch of the dataloader,
             # make sure we feed a full batch of noise
-            z = torch.randn((opt.batchSize, nz)).view(-1, nz, 1, 1)
-            z = Variable(z.cuda()) if opt.cuda else Variable(z)
-
-            Gz = Variable(netG(z).data, volatile=True)
-
-            y_fake = Variable(torch.zeros(batch_size))
-            if opt.cuda:
-                real_cpu = real_cpu.cuda()
-                # real_data = real_data.cuda()
-                y_real = y_real.cuda()
-                y_fake = y_fake.cuda()
-
-            # print("-" * 5 + "Train D with real " + "-" * 5)
-            X = Variable(input.resize_as_(real_cpu).copy_(real_cpu))
-            V, reconstructions, masked = netD(X)
-            Vz, reconstructions_z, masked_z = netD(Gz)
-
-            netD_real_loss = netD.loss(data=X, x=V, target=y_real, reconstructions=reconstructions)
-            netD_fake_loss = netD.loss(data=Gz, x=Vz, target=y_fake, reconstructions=reconstructions_z)
-            netD_train_loss = netD_real_loss + netD_fake_loss
-            # print(f"\n\tepoch_{epoch}_batch_{i}_D_real_loss = {netD_real_loss}")
-            # print(f"\tepoch_{epoch}_batch_{i}_D_fake_loss = {netD_fake_loss}\n")
-            print(f"\n\tepoch_{epoch}_batch_{i}_D_train_loss = {netD_train_loss}\n")
-
-            netD_train_loss.backward()
-            optimizerD.step()
-
-            ## Train Generator ##
-            netG.zero_grad()
-            optimizerG.zero_grad()
-            netD.eval()
-            z = torch.randn((opt.batchSize, nz)).view(-1, nz, 1, 1)
-            z = Variable(z.cuda()) if opt.cuda else Variable(z)
-            Gz = Variable(netG(z).data, volatile=True)
-            y_real = Variable(torch.ones(batch_size))
-            if opt.cuda:
-                y_real = y_real.cuda()
-            Vz, reconstructions_z, masked_z = netD(Gz)
-            netG_train_loss = netD.loss(data=Gz, x=Vz, target=y_real, reconstructions=reconstructions_z)
-            print(f"\n\tepoch_{epoch}_batch_{i}_G_train_loss = {netG_train_loss}\n")
-            netG_train_loss.backward()
+            noise.resize_(opt.batchSize, nz, 1, 1).normal_(0, 1)
+            noisev = Variable(noise)
+            fake = netG(noisev)
+            errG = netD(fake)
+            errG.backward(one)
             optimizerG.step()
+            gen_iterations += 1
 
-            # gen_iterations += 1
-            # print("\tgen_iterations: %d" % gen_iterations)
+            print('[%d/%d][%d/%d][%d] Loss_D: %f Loss_G: %f Loss_D_real: %f Loss_D_fake %f'
+                  % (epoch, opt.niter, i, len(dataloader), gen_iterations,
+                     errD.data[0], errG.data[0], errD_real.data[0], errD_fake.data[0]))
 
             if visualize:
-                netD_loss_logger.log(epoch, netD_train_loss)
-                netG_loss_logger.log(epoch, netG_train_loss)
+                netD_loss_logger.log(epoch, errD.data[0])
+                netD_loss_logger.log(epoch, errG.data[0])
 
-            # if gen_iterations % 10 == 0 or ((gen_iterations % 100 == 0) and (opt.dataset == 'mnist')):
-            if i % 10 == 0 or ((i % 100 == 0) and (opt.dataset == 'mnist')):
-                real_cpu = real_cpu.mul(0.5).add(0.5)       #de-normalizing mnist to get real sample x_real = mu + sigma.z
+            if gen_iterations % 500 == 0 or ((gen_iterations % 100 == 0) and (opt.dataset == 'mnist')):
+                real_cpu = real_cpu.mul(0.5).add(0.5)
                 vutils.save_image(real_cpu, '{0}/{1}/real_samples.png'.format(opt.experiment, opt.dataset))
-                fake = netG(Variable(fixed_noise, volatile=True))   # Get a sample from random data from the generator
+                fake = netG(Variable(fixed_noise, volatile=True))
                 fake.data = fake.data.mul(0.5).add(0.5)
-                vutils.save_image(fake.data, '{0}/{1}/fake_samples_epoch_{2}_batch_{3}.png'.format(opt.experiment, opt.dataset, epoch, i))
+                vutils.save_image(fake.data, '{0}/{1}/fake_samples_{2}.png'.format(opt.experiment, opt.dataset, gen_iterations))
 
         # do checkpointing
         if opt.niter > 25:
